Create opening_clock_with_word.py
Libraries used: OS, PySimpleGUI, datetime/time, sys

What this program does is check for the existence of Microsoft Word, that is to say, checks the list of running processes for Word. If Word is found to be open, open a persistent window in the center and upper part of the screen. This window shows the clock (in 24hrs) and elapsed time (in H:M format). When Word is closed, close the window automatically.